Reorganize CSS files structure
- Moved all CSS modules from src/ to src/styles/components/
- Updated import paths in Contact.jsx and PrelectionsAndCourses.jsx
- Removed duplicate CSS files from src/
- All CSS files now properly organized in styles/ directory structure
- Verified application compiles successfully
