--- frontend/src/App.module.css
@@ -1,174 +1,0 @@
-/* landingpage/src/App.module.css */
-
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@700&family=Indie+Flower&family=Playfair+Display:wght@400;700&display=swap');
-
-:root {
-  --color-bg1: #181c2b;
-  --color-bg2: #232946;
-  --color-accent: #ffb800;
-  --color-text: #fff;
-  --color-subtext: #c9c9c9;
-}
-
-.app-container {
-  color: var(--color-text);
-  font-family: 'Montserrat', Arial, sans-serif;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-main {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-}
-
-.hero-viewport {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  background-color: #181c2b; /* Fallback for when image fails */
-}
-
-.page-content {
-  flex: 1;
-}
-
-/* .navbar {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 2rem 3rem 1rem 3rem;
-  font-size: 1.1rem;
-  background: transparent;
-}
-
-.navbar__logo {
-  font-weight: bold;
-  letter-spacing: 2px;
-  font-size: 1.3rem;
-  color: var(--color-accent);
-}
-
-.navbar__links {
-  display: flex;
-  gap: 2.5rem;
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.navbar__link {
-  color: var(--color-text);
-  text-decoration: none;
-  transition: color 0.2s;
-  position: relative;
-}
-
-.navbar__link:hover {
-  color: var(--color-accent);
-} */
-
-.main-content {
-  flex: 1; /* Allow main content to grow and push gallery down */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 5rem;
-  padding: 2rem 6vw;
-  position: relative;
-}
-
-.hero {
-  max-width: 50vw;
-  z-index: 2;
-}
-
-.hero__headline {
-  font-size: 2.2rem;
-  font-weight: 700;
-  margin-bottom: 1.2rem;
-  line-height: 1.1;
-  color: var(--color-text);
-  text-shadow: 0 2px 16px #000a;
-}
-
-.hero__sub {
-  margin-bottom: 2.5rem;
-}
-
-.hero__signature {
-  margin-top: 1.5rem;
-}
-
-.hero__name {
-  font-family: 'Playfair Display', serif;
-  font-weight: 700;
-  font-size: 1.5rem;
-  color: var(--color-accent);
-  margin-bottom: 0.3rem;
-  letter-spacing: 0.5px;
-}
-
-.hero__title {
-  font-family: 'Montserrat', sans-serif;
-  font-size: 0.9rem;
-  color: var(--color-subtext);
-  text-transform: uppercase;
-  letter-spacing: 2px;
-}
-
-.hero__stars {
-  font-size: 1.2rem;
-  margin: 0 0.3rem;
-  vertical-align: middle;
-}
-
-.astro-image {
-  width: 280px;
-  height: 280px;
-  border-radius: 50%;
-  object-fit: cover;
-  box-shadow: 0 8px 32px #000a;
-  background: #222;
-  display: block;
-  margin-left: 2rem;
-}
-
-@media (max-width: 900px) {
-  .main-content {
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 2rem 4vw;
-  }
-  .hero {
-    max-width: 100vw;
-  }
-  .astro-image {
-    margin: 2rem auto 0 auto;
-    width: 200px;
-    height: 200px;
-  }
-}
-
-.footer {
-  background: transparent;
-  color: var(--color-subtext);
-  text-align: center;
-  padding: 1rem 0 0.5rem 0;
-  font-size: 0.95rem;
-}
-
-.loading-indicator, .error-message {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-  font-size: 1.5rem;
-}
-
-.error-message {
-  color: #ff6b6b;
-}
