--- frontend/src/AstroGallery.module.css
@@ -1,199 +1,0 @@
-.container {
-  text-align: center;
-  background-color: #0f172a; /* Dark blue-gray background */
-  min-height: 100vh;
-  color: #e0e0e0;
-}
-
-.title {
-  font-size: 3rem;
-  color: #ffffff;
-  margin-bottom: 3rem;
-  font-weight: 300;
-  letter-spacing: 2px;
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(3, 350px);
-  gap: 1.5rem;
-  max-width: calc(3 * 350px + 2 * 1.5rem);
-  margin: 0 auto;
-  justify-content: center;
-}
-
-.gridItem {
-  width: 350px;
-  height: 350px;
-  background-color: #1e293b;
-  border-radius: 8px;
-  overflow: hidden;
-  box-shadow: 0 4px 8px rgba(0,0,0,0.2);
-  transition: transform 0.3s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.gridItem:hover {
-  transform: scale(1.03);
-}
-
-.gridItem img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  display: block;
-}
-
-.placeholder {
-  font-size: 1.2rem;
-}
-
-.hero {
-  width: 100vw;
-  height: 33vh;
-  min-height: 250px;
-  background-size: cover;
-  background-position: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 2rem;
-  position: relative;
-  padding-top: 80px; /* Add space for navbar overlay */
-}
-
-.hero::after {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(15, 23, 42, 0.5); /* dark overlay for readability */
-  z-index: 1;
-}
-
-.heroTitle {
-  color: #fff;
-  font-size: 3rem;
-  font-weight: 700;
-  letter-spacing: 2px;
-  z-index: 2;
-  position: relative;
-}
-
-.filtersSection {
-  display: flex;
-  justify-content: center;
-  gap: 1rem;
-  margin: 1.5rem 0 2.5rem 0;
-  flex-wrap: wrap;
-}
-
-.filterBox {
-  background: #1e293b;
-  color: #fff;
-  border-radius: 8px;
-  padding: 0.5rem 1.2rem;
-  font-size: 1rem;
-  font-weight: 500;
-  cursor: pointer;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  transition: background 0.2s, color 0.2s, transform 0.2s, border-color 0.2s;
-  border: 2px solid transparent;
-}
-
-.filterBox:not(.activeFilter):hover {
-  background: #334155;
-  color: #ffd700;
-  border-color: #ffd700;
-}
-
-.activeFilter {
-  background: #ffd700 !important;
-  color: #181c2b !important;
-  border-color: #ffd700 !important;
-  font-weight: 700;
-  box-shadow: 0 4px 16px rgba(255,215,0,0.12);
-}
-
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background: rgba(0,0,0,0.7);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1000;
-  overflow-y: auto; /* Allow scrolling if modal is too tall */
-}
-
-.modalContent {
-  background: #181c2b;
-  border-radius: 12px;
-  padding: 2rem 2rem 1.5rem 2rem;
-  box-shadow: 0 8px 32px rgba(0,0,0,0.4);
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-width: 320px;
-  max-width: 1000px;
-  width: 90vw;
-  margin: 5vh 0; /* Add vertical margin for top and bottom space */
-  max-height: 90vh; /* Prevent modal from exceeding viewport */
-  overflow-y: auto; /* Make modal scrollable if content is too tall */
-  -ms-overflow-style: none;  /* IE and Edge */
-  scrollbar-width: none;     /* Firefox */
-}
-
-.modalContent::-webkit-scrollbar {
-  display: none;
-}
-
-.modalImage {
-  max-width: 90%;
-  border-radius: 8px;
-  margin-bottom: 1.5rem;
-  object-fit: contain;
-  display: block;
-}
-
-.modalDescription {
-  color: #fff;
-  font-size: 1.1rem;
-  text-align: center;
-  margin-top: 0.5rem;
-  font-weight: 400;
-  font-family: 'Roboto', Arial, Helvetica, sans-serif;
-  padding: 1.5rem 2rem;
-  background: rgba(24, 28, 43, 0.85);
-  border-radius: 8px;
-  box-sizing: border-box;
-  word-break: break-word;
-  white-space: pre-line;
-  max-width: 100%;
-  overflow-wrap: break-word;
-}
-
-.modalClose {
-  position: absolute;
-  top: 0.5rem;
-  right: 1rem;
-  background: none;
-  border: none;
-  color: #fff;
-  font-size: 2rem;
-  cursor: pointer;
-  z-index: 10;
-  transition: color 0.2s;
-}
-
-.modalClose:hover {
-  color: #ffd700;
-}
