--- frontend/src/Navbar.module.css
@@ -1,79 +1,0 @@
-/* landingpage/src/Navbar.module.css */
-
-:local(.navbar) {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1.2rem 2rem;
-  background: transparent;
-  color: #fff;
-  box-sizing: border-box;
-  overflow-x: visible;
-}
-
-:local(.navbar__logo) {
-  height: 100px;
-  background: none;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.10);
-  filter: drop-shadow(0 2px 8px rgba(0,0,0,0.25)); /* subtle shadow for visibility */
-  display: block;
-  opacity: 50%;
-}
-
-:local(.navbar__links) {
-  list-style: none;
-  display: flex;
-  gap: 2rem;
-  margin: 0;
-  padding: 0;
-  font-family: 'Montserrat', Arial, sans-serif;
-}
-
-:local(.navbar__link) {
-  color: #fff;
-  text-decoration: none;
-  font-size: 1.1rem;
-  transition: color 0.2s;
-  font-family: 'Montserrat', Arial, sans-serif;
-}
-
-:local(.navbar__link:hover) {
-  color: #9a8c98;
-}
-
-@media (max-width: 600px) {
-  :local(.navbar) {
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 1rem;
-  }
-  :local(.navbar__links) {
-    gap: 1rem;
-    flex-direction: column;
-    width: 100%;
-  }
-} 
-
-.transparent {
-  background: transparent !important;
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  z-index: 10;
-  color: #fff;
-  box-sizing: border-box;
-}
-
-.transparent a,
-.transparent a:visited,
-.transparent a:active,
-.transparent ul li a {
-  color: #fff !important;
-  text-shadow: 0 1px 4px rgba(0,0,0,0.3);
-}
-
-.transparent ul li a:hover,
-.transparent ul li a:focus {
-  color: #ffd700 !important;
-}
