+++ frontend/src/styles/components/PrelectionsAndCourses.module.css
